+++ srp-conversion-support/src/main/java/org/srplib/conversion/TypeAwareConverter.java
@@ -1,0 +1,18 @@
+package org.srplib.conversion;
+
+/**
+ * Source and target type aware converter.
+ */
+public interface TypeAwareConverter<I, O> extends Converter<I, O> {
+
+    /**
+     * Source type.
+     */
+    Class<I> getInputType();
+
+    /**
+     * Target type.
+     */
+    Class<O> getOutputType();
+
+}